+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/components/SessionQueueTable.scala
@@ -21,7 +21,7 @@ import monocle.macros.Lenses
 import react.virtualized._
 import scala.scalajs.js
 import scala.math.max
-import react.common._
+import react.common.{Css, Size, Style}
 import react.common.implicits._
 import seqexec.model.enum.Instrument
 import seqexec.model.UserDetails
@@ -643,7 +643,7 @@ object SessionQueueTable extends Columns {
   }
 
   def table(b: Backend)(size: Size): VdomNode =
-    if (size.width > 0) {
+    if (size.width.toInt > 0) {
       Table(
         Table.props(
           disableHeader = false,
